+++ app/pages/invoices/pay_invoice/script.js
@@ -1,19 +1,28 @@
 ﻿$("document").ready(function () {
 
-    // Get the token
-    var token = localStorage.getItem("token");
-
-    // Define the host
-    var host = "api.comecero.com";
-    if (window.location.hostname.indexOf("-staging.") > -1) {
-        host = "api-staging.comecero.com";
+    function getCookie(name) {
+        if (document.cookie.length > 0) {
+            c_start = document.cookie.indexOf(name + "=");
+            if (c_start != -1) {
+                c_start = c_start + name.length + 1;
+                c_end = document.cookie.indexOf(";", c_start);
+                if (c_end == -1) {
+                    c_end = document.cookie.length;
+                }
+                return unescape(document.cookie.substring(c_start, c_end));
+            }
+        }
+        return "";
     }
+
+    // Get the token
+    var token = getCookie("token");
 
     // Get the query parameters
     var params = utils.getPageQueryParameters();
 
     // Define the URL
-    var url = "https://" + host + "/api/v1/invoices/" + params["invoice_id"] + "/links";
+    var url = "/api/v1/invoices/" + params["invoice_id"] + "/links";
 
     // Make a request to get the notification body
     if (params["invoice_id"]) {