+++ app/shared/apiServices.js
@@ -17,17 +17,21 @@
             data = undefined;
         }
 
+        // Remove any existing token in storage
+        utils.deleteCookie("token");
+
         var headers = {};
         headers["Content-Type"] = "application/json";
 
         var request = $http({
-            ignoreLoadingBar: true,
+            ignoreLoadingBar: false,
             method: "post",
             data: angular.toJson(data),
             url: url + "?timezone=UTC",
             params: parameters,
             timeout: 15000,
             isApi: true,
+            isLogin: true,
             ignoreAuthModule: true,
             headers: headers
         });
@@ -227,13 +231,6 @@
 
         // Most calls will return a formatted error message unless something unexpected happens. Pass the error object through if present, or create one that has the same properties if not.
         if (response.data.error) {
-
-            if (response.data.error.status == 403) {
-                error.code = "error";
-                error.message = "It appears that you don't have permissions to access page or function you have requested. If you feel this is incorrect, please contact an account administrator.";
-                error.status = response.status;
-                return ($q.reject(error));
-            }
 
             return ($q.reject(response.data.error));