+++ app/shared/commonServices.js
@@ -108,8 +108,22 @@ app.service("SettingsService", ['$rootScope', "$q", "ApiService", function ($roo
             return appSettings;
         }
 
+        // Get style settings
+        var getStyleSettings = function () {
+
+            var styleSettings = {};
+
+            if (window.__settings) {
+                if (window.__settings.style) {
+                    styleSettings = window.__settings.style;
+                }
+            }
+
+            return styleSettings;
+        };
+
         // Build and return the settings object
-        var settings = { account: getAccountSettings(), app: getAppSettings(), config: {} };
+        var settings = { account: getAccountSettings(), app: getAppSettings(), style: getStyleSettings(), config: {} };
 
         // Define the api prefix
         settings.config.apiPrefix = "/api/v1";
@@ -121,11 +135,9 @@ app.service("SettingsService", ['$rootScope', "$q", "ApiService", function ($roo
 
             settings.config.development = true;
 
-            var apiHost = settings.account.api_host || settings.app.api_host || settings.style.api_host || "api.comecero.com";
-            apiHost = "https://" + apiHost;
-
             // Make the apiPrefix a fully qualified url since requests in development mode don't have access to the reverse proxy.
-            settings.config.apiPrefix = apiHost + settings.config.apiPrefix;
+            var apiHost = settings.account.api_host || settings.app.api_host || "api.comecero.com";
+            settings.config.apiPrefix = ("https://" + apiHost) + settings.config.apiPrefix;
         }
 
         return settings;