+++ settings/style-SAMPLE.js
@@ -1,0 +1,13 @@
+﻿// Make a copy of this file and name it style.js, which can be used by the app to customize app styles.
+// Note that the style values provided by the user are compiled into a css stylesheet (located in sass/style.css). This file gives you access to the raw user style input values.
+// At runtime, the user's style settings will be automatically injected, providing the user's selections to the app.
+window.__settings = window.__settings || {}; window.__settings.style =
+    {
+        "sidebar-color": "#2c3e50",
+        "sidebar-text": "#f6f7f8",
+        "logo_medium": null,
+        "logo_large": null,
+        "logo_square": null,
+        "logo_full": null,
+        "logo": null,
+    };