+++ CHANGELOG.md
@@ -1,9 +1,24 @@
-﻿<a name="1.0.1"></a>
+﻿<a name="1.1.1"></a>
+# 1.1.1
+
+- Hide subscription item edit button when customer subscription downgrades and upgrades are not permitted
+
+
+<a name="1.1.0"></a>
+# 1.1.0
+
+- Refactor login page
+- Allow custom colors, logo, favicon
+- Add support for subscription modifications
+- Bug fixes and UI refactors
+
+<a name="1.0.1"></a>
 # 1.0.1
 
 - Replace dependency on third party CDNs and replace with internal CDN.
 - Update to HTML references so the app can run from outside of the root URL.
 - Numerous minor bug fixes.
+
 
 <a name="1.0.0"></a>
 # 1.0.0